refactor markdown building to separate file
